Added project board resource, service.

+++ src/main/java/com/mbancer/web/rest/dto/BoardDTO.java
@@ -4,15 +4,15 @@ import java.util.Objects;
 
 public class BoardDTO {
 
-    public int id;
+    public Long id;
     public String name;
-    public int projectId;
+    public Long projectId;
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
@@ -24,11 +24,11 @@ public class BoardDTO {
         this.name = name;
     }
 
-    public int getProjectId() {
+    public Long getProjectId() {
         return projectId;
     }
 
-    public void setProjectId(int projectId) {
+    public void setProjectId(Long projectId) {
         this.projectId = projectId;
     }
